refactor: manage packages and names (#7)
