+++ internal/service/service.go
@@ -1,25 +1,12 @@
-package utils
+// Package service is helper package for services which is external data and need to be parsed
+// to provide usefull information such as (port number, protocol and service name)
+package service
 
 import (
 	"encoding/csv"
 	"fmt"
-	"net"
 	"os"
 )
-
-// Listen on port 0 to get a free port assigned by the system.
-// Get the actual address, including the assigned port.
-func GetFreeTCPPort() (int, error) {
-
-	listener, err := net.Listen("tcp", ":0")
-	if err != nil {
-		return 0, err
-	}
-	defer listener.Close()
-
-	addr := listener.Addr().(*net.TCPAddr)
-	return addr.Port, nil
-}
 
 // ServiceInfo represents information about a service
 type ServiceInfo struct {
