--- utils/doc.go
@@ -1,2 +1,0 @@
-// Package utils offer a set of helpers for the scanner
-package utils
